split MINIT for CharacterDirection
