+++ package.xml
@@ -83,6 +83,8 @@ Fix: ensure build directories are created for each source directory, fixing an i
             <file name="locale.h" role="src"/>
             <file name="locale_character_direction.c" role="src"/>
             <file name="locale_character_direction.h" role="src"/>
+            <file name="locale_character_direction_arginfo.h" role="src"/>
+            <file name="locale_character_direction.stub.php" role="src"/>
             <file name="locale_options.c" role="src"/>
             <file name="locale_options.h" role="src"/>
             <file name="locale_options_arginfo.h" role="src"/>
